# swagger/swagger-dev.yaml
swagger: "2.0"
info:
  title: MyECP/MMA Swagger
  description: Middleware API on API Gateway with a Google Cloud Functions backend
  version: 1.0.0
schemes:
  - https
host: apigateway-dev-ccoox6j1.uc.gateway.dev
basePath: /api
...
produces:
  - application/json
paths:
...
  /cards/validate-security-questions:
    options:
      tags:
        - Apply Card
      operationId: corsApplyCardValidateSecurityQuestions
      x-google-backend:
        address: https://us-central1-myecpmma-non-prod.cloudfunctions.net/ApplyCardValidateSecurityQuestions
      responses:
        200:
          description: Success

    post:
      tags:
        - Apply Card
      operationId: ApplyCardValidateSecurityQuestions
      consumes:
        - application/json
      produces:
        - application/json
      x-google-backend:
        address: https://us-central1-myecpmma-non-prod.cloudfunctions.net/ApplyCardValidateSecurityQuestions
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/ApplyCardValidateSecurityQuestionsRequest"
      responses:
        405:
          description: "Invalid input"
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/ApplyCardStatusResponse"
        500:
          description: "Internal Server Error"
        401:
          description: "Unauthorized Users"

  
          
  # Offers
...
definitions:
...
  ApplyCardValidateSecurityQuestionsRequest:
    type: object
    properties:
      ID:
        type: integer
      UserId:
        type: integer
      SecurityQuestions:
        type: "array"
        items: { 
          $ref: "#/definitions/ApplyCardSecurityQuestions" 
        }      
...
  ApplyCardSecurityQuestions:
    type: object
    properties:
      QuestionID:
        type: integer
      Question:
        type: string
...
  ApplyCardStatusResponse:
    type: object
    properties:
      ApplicationNumber:
        type: number
        example: 20247986465874560
      ApplicationType:
        type: string
        example: Military Clothing
...